refactor: adjust namespaces
