+++ DirectoryService/src/DirectoryService.Contracts/Requests/PositionRequests/AddPositionRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace DirectoryService.Contracts.Requests;
+﻿namespace DirectoryService.Contracts.Requests.PositionRequests;
 
 public class AddPositionRequest
 {
